working finding challenges by id and removing challenge when all challenged users decline challenge

+++ src/main/java/com/capgemini/challenges/challenge/service/ChallengeService.java
@@ -1,10 +1,12 @@
 package com.capgemini.challenges.challenge.service;
 
 import com.capgemini.challenges.challenge.Challenge;
+import com.capgemini.challenges.challenge.UserStatus;
 import com.capgemini.challenges.challenge.dao.ChallengeDAO;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import java.util.Date;
+import java.util.List;
 
 public class ChallengeService {
     private ChallengeDAO challengeDAO;
@@ -15,30 +17,55 @@ public class ChallengeService {
     }
 
     //jakis validator ze playerOneId nie moze byc rowny z playerTwoId
-    public void createChallenge(long gameId, long playerOneId, long plyerTwoId){
+    public void createChallenge(long gameId, List<Long> playersId) {
         Challenge challenge = new Challenge();
         challenge.setGameId(gameId);
-        challenge.setPlayerOneId(playerOneId);
-        challenge.setPlayerTwoId(plyerTwoId);
+        challenge.setPlayersId(playersId);
         challenge.setDateOfChallenge(new Date());
         challenge.setGameStatus(false);
-        challenge.setStatusPlayerOne(false);
-        challenge.setStatusPlayerTwo(false);
         challengeDAO.addChallenge(challenge);
     }
 
-
-    //beda potrzebne validatory czy challenge jest konkretnego usera i tylko on moze odrzucic lub przyjac
-    public void acceptChallenge(long challengeId){
+    //usunac duplikaty pozniej
+    //beda potrzebne validatory czy challenge jest konkretnego usera i tylko on moze odrzucic lub przyjac, jesli przyjma dwie osoby to reszta nie moze
+    public void acceptChallenge(long playerId, long challengeId) {
         Challenge challenge = challengeDAO.findChallengeById(challengeId);
-        challengeDAO.removeChallenge(challenge);
-        //jeszcze jakos obczaic ktory player jest ktory
-        challenge.setStatusPlayerOne(true);
-        challengeDAO.addChallenge(challenge);
+        int playerNumber = challenge.getPlayersId().indexOf(playerId);
+        List<UserStatus> statuses = challenge.getStatusesOfPlayers();
+        statuses.set(playerNumber, UserStatus.ACCEPTED);
+        challenge.setStatusesOfPlayers(statuses);
+        //challengeDAO.addChallenge(challenge);
     }
 
-    public void declineChallenge(long challengeId){
+    //tutaj by sie przdalo sprawdzic czy wszyscy odrzucili challenge, jak tak to usunac
+    public void declineChallenge(long playerId, long challengeId) {
         Challenge challenge = challengeDAO.findChallengeById(challengeId);
-        challengeDAO.removeChallenge(challenge);
+        int playerNumber = challenge.getPlayersId().indexOf(playerId);
+        List<UserStatus> statuses = challenge.getStatusesOfPlayers();
+        statuses.set(playerNumber, UserStatus.DECLINED);
+
+        //to pojdzie do osobnej metody i zrobic streama z tego
+        int countDeclined = 0;
+        for (UserStatus status : challenge.getStatusesOfPlayers()) {
+            if (!status.equals(UserStatus.DECLINED)) {
+                break;
+            } else {
+                countDeclined++;
+                continue;
+            }
+        }
+
+        if(countDeclined == challenge.getStatusesOfPlayers().size()){
+            challengeDAO.removeChallenge(challenge);
+        }
+    }
+
+    public List<Challenge> showAcceptedChallenges(long playerId) {
+        return null;
+    }
+
+    //tutaj dodawanie pkt zwyciezcy
+    public void endOfChallenge() {
+
     }
 }
